ai-blog-generator/requirements.txt
Flask==2.3.2
openai==1.84.0
python-dotenv==1.0.0
apscheduler==3.10.1
requests==2.32.0
pytz~=2025.2